requirements.txt
appnope==0.1.2
asttokens==2.0.5
autograd==1.3
backcall==0.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
celluloid==0.2.0
cma==2.7.0
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fonttools==4.31.2
future==0.18.2
ipykernel==6.10.0
ipython==8.2.0
jedi==0.18.1
jupyter-client==7.2.0
jupyter-core==4.9.2
kiwisolver==1.4.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
numpy==1.22.3
openpyxl==3.0.9
packaging==21.3
pandas==1.4.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prompt-toolkit==3.0.28
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pygad==2.16.3
Pygments==2.11.2
pymoo==0.5.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
scipy==1.8.0
six==1.16.0
soupsieve==2.3.1
stack-data==0.2.0
tornado==6.1
traitlets==5.1.1
wcwidth==0.2.5
